% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{availableCrops}
\alias{availableCrops}
\title{Get crop names available for a given location.}
\usage{
availableCrops(lat, lng, loc.data = locationCrops, crop.data = crops)
}
\arguments{
\item{lat}{the latitude.}

\item{lng}{the longitude.}
}
\value{
vector of crop names.
}
\description{
Given a lat/lng pair, gets a list of cropids and then
fetches crop names based on those ids.
}

