% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{averageTemperatures}
\alias{averageTemperatures}
\title{Average temperatures over multiple years.}
\usage{
averageTemperatures(df)
}
\arguments{
\item{df}{a data frame of climate data for multiple years.}
}
\value{
a dataframe consisting of a single year's worth of average data.
}
\description{
Convert temperature data for multiple years into a single average,
with accompanying measures like standard deviation, percentages, etc...
}

