% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculators.R
\name{calcCumGDD}
\alias{calcCumGDD}
\title{calculate cumulative GDD}
\usage{
calcCumGDD(dates, temp.hi, temp.low, ...)
}
\arguments{
\item{dates}{a vector of formatted dates}

\item{temp.hi}{a vector of daily high temperatures}

\item{temp.low}{a vector of daily low temperatures}

\item{base}{a vector of the crop-dependent base temperature}
}
\value{
a vector of cumulative GDD
}
\description{
calculates cumulative Growing Degree Days
}

