% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculators.R
\name{calcGDD}
\alias{calcGDD}
\title{Calculate growing degree days.}
\usage{
calcGDD(temp.hi, temp.low, temp.base = 50, restrict = 25, temp.max = 86,
  temp.min = 50)
}
\arguments{
\item{temp.hi}{a vector of daily high temperatures}

\item{temp.low}{a vector of daily low temperatures}

\item{temp.max}{a number denoting maximum temperature (default=86)}

\item{temp.min}{a number denoting minimum temperature (default=50)}

\item{base}{a vector of the crop-dependent base temperature}
}
\value{
a vector of (noncumulative) GDD
}
\description{
Calculates Growing Degree Days from temperature data.
}

