% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cbccy.R
\docType{package}
\name{cbccy}
\alias{cbccy}
\alias{cbccy-package}
\title{cbccy-package: Tools and analysis for the impact of climate change on the Columbia River Basin}
\description{
This package contains the code and various implementations of decision-tools used to evaluate the impact of climate change on crop yields
}
\author{
Nicholas Potter
}
\keyword{cbccy-package}
\keyword{package}

