% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{createChartData}
\alias{createChartData}
\title{makePlot() creates a plot for the app}
\usage{
createChartData(df, this.year)
}
\arguments{
\item{df}{a data frame}
}
\description{
generate an interactive plot using ggplot2.
}

