% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{createDailyData}
\alias{createDailyData}
\title{Create daily range climate data.}
\usage{
createDailyData(df)
}
\arguments{
\item{df}{a data frame of climate data spanning multiple years.}
}
\value{
a data frame containing growing degree days.
}
\description{
Create a data frame with max/min temp, precip, windspeed, gdd, etc....
}

