% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{createPlotFromClimateData}
\alias{createPlotFromClimateData}
\title{make the central plot}
\usage{
createPlotFromClimateData(df.historical, df.present, df.future = NULL, crop,
  today = format(Sys.Date(), "\%Y-\%m-\%d"))
}
\arguments{
\item{df}{a data frame containing gdd variables.}

\item{gdd.silk}{(numeric) the gdd level needed to silk (specific to a crop)}

\item{gdd.flower}{(numeric) the gdd level needed to flower (specific to a crop)}
}
\value{
a plot object made with ggplot2
}
\description{
Plots the climate data as well as yield estimates and growth stage points.
}

