% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{createPlotFromLocation}
\alias{createPlotFromLocation}
\title{Create a plot for a specific location from scratch.}
\usage{
createPlotFromLocation(lat, lng, crop, years.this, years.range = c(1970,
  2006), years.compare = NULL, today = Sys.Date(), forecast.date = NULL)
}
\arguments{
\item{lat}{the latitude of a location.}

\item{lng}{the longitude of a location.}

\item{crop}{the name of a crop.}

\item{years.range}{a 2-item list of the first and last
year to include in the range.}

\item{years.compare}{list of years to compare to.}

\item{today}{the date to consider today.}

\item{forecast.date}{the data at which forecast starts.}

\item{year.this}{the year to consider the current year.}
}
\value{
plot object
}
\description{
Loads the necessary data and creates a plot from it.
}

