% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{dateVector}
\alias{dateVector}
\title{Create a vector of dates.}
\usage{
dateVector(year.start = 1915, year.end = 2006)
}
\arguments{
\item{year.start}{the year to start with, default 1915}

\item{year.end}{the last year, default 2006}
}
\value{
vector of date-formatted values
}
\description{
Creates a vector of dates from a start and end year, inclusive.
}

