% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{fetchRawData}
\alias{fetchRawData}
\title{Fetch raw data}
\usage{
fetchRawData(username = NULL, password = NULL, server = "aeolus.wsu.edu",
  save.dir = "data/", file.paths = c("/data/kirti/",
  "/data/jennylabcommon/Future2035MetData/",
  "/data/jennylabcommon/Future2035MetData/",
  "/data/jennylabcommon/Future2035MetData/",
  "/data/jennylabcommon/Future2035MetData/",
  "/data/jennylabcommon/Future2035MetData/"),
  file.names = c("vic_inputdata0625_pnw_combined_05142008.tar.gz",
  "ccsm3_B1_2020-2049.tar.gz", "cgcm3.1_t47_B1_2020-2049.tar.gz",
  "hadcm_B1_2020-2049.tar.gz", "ipsl_cm4_A1B_2020-2049.tar.gz",
  "pcm1_A1B_2020-2049.tar.gz"))
}
\arguments{
\item{username}{login username}

\item{password}{login password}

\item{server}{server name, default aeolus.wsu.edu}

\item{save.dir}{directory to save data to}

\item{file.paths}{list of file paths}

\item{file.names}{list of file names to download}
}
\value{
TRUE if successful
}
\description{
Get raw data from WSU servers and store it in data/
}

