% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{getClimateData}
\alias{getClimateData}
\title{Load climate data.}
\usage{
getClimateData(lat, lng, data.dir = "data/locations/",
  file.name.root = "data_", column.names = c("precipitation", "temp.max",
  "temp.min", "windspeed"), column.factors = c(0.4, 0.01, 0.01, 0.01))
}
\arguments{
\item{lat}{a latitude}

\item{lng}{a longitude}

\item{data.dir}{a directory location, default "data/locations/"}

\item{file.name.root}{filename roots, default "data_"}

\item{column.names}{data frame column names, default
c("precipitation", "temp.max", "temp.min", "windspeed")}

\item{column.factors}{a vector of values to multiply by}
}
\value{
a data frame of climate data
}
\description{
Loads climate data for a specific latitude/longitude pair.
}

