% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{getData}
\alias{getData}
\title{Get chart data for a location.}
\usage{
getData(year.start, year.end)
}
\arguments{
\item{year.start}{the first year of data to include}

\item{year.end}{the last year of data to include}
}
\value{
data frame of data for graphic
}
\description{
Load climate data and calculate values to provide chart data.
}

