% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{getHistoryData}
\alias{getHistoryData}
\title{Get data from historical file}
\usage{
getHistoryData(lat, lng, year.start = 1915, year.end = 2006, ...)
}
\arguments{
\item{lat}{the latitude of a location}

\item{lng}{the longitude of a location}

\item{year.start}{the first year of data to include, default 1915}

\item{year.end}{the last year of data to include, default 2006}
}
\value{
a data frame of climate data from 1915-2006
}
\description{
Gets data from history, including specific years of history
}

