% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{getLocationCrops}
\alias{getLocationCrops}
\title{Get crops available for a given location.}
\usage{
getLocationCrops(lat, lng, data.dir = "data/")
}
\arguments{
\item{lat}{the latitude.}

\item{lng}{the longitude.}

\item{data.dir}{the directory of the crop lookup data.}
}
\value{
a list of cropids.
}
\description{
Given a lat/lng pair, lookup which crop ids are produced
in that area. Return the cropids.
}

