% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{getLocationData}
\alias{getLocationData}
\title{Get climate data for a location.}
\usage{
getLocationData(lat, lng, mode, data.dir = "data/", ...)
}
\arguments{
\item{lat}{the latitude of a location}

\item{lng}{the longitude of a location}

\item{mode}{one of 'historical', 'future', or 'present'.
Determines data to fetch.}

\item{data.dir}{the directory containing 'historical', 'future',
and 'present' directories.}

\item{...}{other parameters passed to loadClimateData()}

\item{crop}{a crop object containing gdd information}
}
\value{
a data frame of climate data with dates attached.
}
\description{
Get climate data for a location. Wraps loadClimateData()
in order to return a data frame with all necessary components
for easy plotting.
}

