% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{makeGDDPlot}
\alias{makeGDDPlot}
\title{make the central plot}
\usage{
makeGDDPlot(dfbyday, dfyear, gdd.silk, gdd.flower)
}
\arguments{
\item{gdd.silk}{(numeric) the gdd level needed to silk (specific to a crop)}

\item{gdd.flower}{(numeric) the gdd level needed to flower (specific to a crop)}

\item{df}{a data frame containing gdd variables.}
}
\value{
a plot object made with ggplot2
}
\description{
Plots the climate data as well as yield estimates and growth stage points.
}

