% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{makePlot}
\alias{makePlot}
\title{Make a multi-plot plot.}
\usage{
makePlot(df)
}
\arguments{
\item{df}{a data frame containing temperature and gdd variables}
}
\value{
a ggplot2 object
}
\description{
Make a multi-plot object including temp and gdd.
}

