% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{makeTempPlot}
\alias{makeTempPlot}
\title{Plot temperature.}
\usage{
makeTempPlot(df, tmax.varname = temp.max)
}
\arguments{
\item{df}{a data frame containing variables: temp.max, temp.min, temp.high, temp.low}
}
\value{
a ggplot2 object
}
\description{
Plot high/low, average, and this year's data
}

