% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{readBinaryData}
\alias{readBinaryData}
\title{Reads binary climate data.}
\usage{
readBinaryData(file.name, what = "integer", size = 2, n = 134412,
  encoding = "ASCII", ...)
}
\arguments{
\item{what}{the data type to read as, default integer}

\item{size}{the byte size, default 2}

\item{n}{the number of entries, default 134412 (92 years)}

\item{encoding}{the binary encoding, default ASCII}

\item{fname}{the name of the file to be read}
}
\value{
vector with read values
}
\description{
Reads binary climate data as a single vector.
}

