% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{selectAndMergeYearData}
\alias{selectAndMergeYearData}
\title{Select and merge year data with day ranges.}
\usage{
selectAndMergeYearData(yeardata, byday, select.years)
}
\arguments{
\item{yeardata}{a df of climate data for range of years.}

\item{byday}{a df of climate data by day.}

\item{select.years}{list of years to plot as comparison.}
}
\value{
a dataframe containing all data necessary to plot a GDD chart.
}
\description{
Select years from year data and merge with day ranges.
}

