% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{setTheme}
\alias{setTheme}
\title{Theme for cbccy plots.}
\usage{
setTheme(p)
}
\arguments{
\item{p}{a ggplot2 object}
}
\value{
a modified ggplot2 object
}
\description{
Sets the default theme for cbccy plots.
}

