% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{writeBinaryData}
\alias{writeBinaryData}
\title{Write a data frame to a binary file.}
\usage{
writeBinaryData(df, file.name, size = 2, encoding = "ASCII")
}
\arguments{
\item{df}{data frame to be written.}

\item{file.name}{the file to write to.}

\item{encoding}{default is ASCII.}
}
\value{
TRUE is successful.
}
\description{
Writes a data frame to binary, with no headers or row names.
The written file is intended to mimic the binary data produced
by the REACH project.
}

