% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{writeClimateData}
\alias{writeClimateData}
\title{Write climate data to a binary file for each location.}
\usage{
writeClimateData(df, file.name, column.names = c("precipitation", "temp.high",
  "temp.low", "windspeed"), column.factors = c(0.025, 0.01, 0.01, 0.01), ...)
}
\arguments{
\item{df}{the data frame to write.}

\item{file.name}{the file name to write to.}

\item{column.names}{the names of columns to write.}

\item{column.factors}{the factors to divide by before writing.}
}
\value{
TRUE if successful.
}
\description{
Writes a climate data frame to a file in binary format.
}

