% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{availableCrops}
\alias{availableCrops}
\title{Get available crops.}
\usage{
availableCrops(dir = "data/locations")
}
\arguments{
\item{dir}{a directory location, default 'data/locations'.}
}
\value{
list of crop
}
\description{
Look at data files to determine what crop data are available.
}

