% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{availableLocations}
\alias{availableLocations}
\title{Get available locations.}
\usage{
availableLocations(dir = "data/locations")
}
\arguments{
\item{dir}{a directory location, default 'data/locations'.}
}
\value{
data frame of latitude/longitude pairs
}
\description{
Look at data files to determine available lat/lng combinations.
}

