% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{fetchRawData}
\alias{fetchRawData}
\title{Fetch raw data}
\usage{
fetchRawData(username = NULL, password = NULL, server = "aeolus.wsu.edu",
  save.dir = "data/",
  file.names = c("/data/kirti/vic_inputdata0625_pnw_combined_05142008.tar.gz",
  "/data/jennylabcommon/Future2035MetData/ccsm3_B1_2020-2049.tar.gz",
  "CropParamsForNick/crops.tar.gz"))
}
\arguments{
\item{username}{login username}

\item{password}{login password}

\item{server}{server name, default aeolus.wsu.edu}

\item{save.dir}{directory to save data to}

\item{file.names}{list of file names to download}

\item{file.paths}{list of file paths}
}
\value{
TRUE if successful
}
\description{
Get raw data from WSU servers and store it in data-raw/
}

