% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{makeGDDPlot}
\alias{makeGDDPlot}
\title{make the central plot}
\usage{
makeGDDPlot(df)
}
\arguments{
\item{df}{a data frame containing the necessary variables}
}
\value{
a plot object made with ggplot2
}
\description{
Plots the climate data as well as yield estimates and growth stage points.
}

